requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-Session
python-dotenv
pytest
pytest-flask
pytest-asyncio
lyricsgenius==3.3.1
torch
transformers
requests
nltk>=3.8
spotipy
psycopg2-binary
Flask-APScheduler
bootstrap-flask
Flask-RQ2
redis
tqdm>=4.65.0
psutil
marshmallow>=3.19.0
cryptography>=40.0.0
prometheus-client>=0.17.0

# Production dependencies
gunicorn[gevent]>=21.2.0
flask-limiter>=3.5.0
prometheus-flask-exporter>=0.23.0
gevent>=23.7.0
